# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Augmented Things",
        "description": "Augmented Things it's a project developed by @Jatapiaro",
        "contact": {
            "email": "jatapiaro@gmail.com"
        },
        "version": "1"
    },
    "paths": {
...
        "/api/v1/places": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Shows the user places",
                "responses": {
                    "200": {
                        "description": "Shows the current user places",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Places"
                ],
                "summary": "Register a new place",
                "requestBody": {
                    "description": "Place that needs to be stored",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "place": {
                                        "type": "object",
                                        "$ref": "#/components/schemas/Place"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Device that was registered.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/api/v1/admin-places": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Shows all the places",
                "responses": {
                    "200": {
                        "description": "Shows all the users places",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/api/v1/places/{place}": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Shows the specified place",
                "parameters": [
                    {
                        "name": "place",
                        "in": "path",
                        "description": "ID of the place",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shows the specified place",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Places"
                ],
                "summary": "Updates a place",
                "parameters": [
                    {
                        "name": "place",
                        "in": "path",
                        "description": "Place to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data of the place to be updated",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "place": {
                                        "type": "object",
                                        "$ref": "#/components/schemas/Place"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Place that was updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Places"
                ],
                "summary": "Deletes a place",
                "parameters": [
                    {
                        "name": "place",
                        "in": "path",
                        "description": "Place to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Place that was deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Place": {
                "title": "Place",
                "description": "Places registered by the user",
                "required": [
                    "name",
                    "latitude",
                    "longitude",
                    "altitude"
                ],
                "properties": {
                    "name": {
                        "schema": "Place",
                        "description": "Name of the place",
                        "type": "string"
                    },
                    "latitude": {
                        "schema": "Place",
                        "description": "Latitude of the place (coordinates)",
                        "type": "number",
                        "example": "19.4284700"
                    },
                    "longitude": {
                        "schema": "Place",
                        "description": "Longitude of the place (coordinates)",
                        "type": "number",
                        "example": "-99.1276600"
                    },
                    "altitude": {
                        "schema": "Place",
                        "description": "Altitude of the place (coordinates)",
                        "type": "number"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}